requirements.txt
PyQt6~=6.7.0
pillow~=11.3.0
pyperclip~=1.11.0
rapidocr~=3.4.1
colorlog~=6.9.0
opencv-python~=4.12.0.88
numpy~=1.26.4
omegaconf~=2.3.0
requests~=2.32.5
tqdm~=4.67.1
torch~=2.8.0
pyclipper~=1.3.0.post6
shapely~=2.1.2
onnxruntime~=1.20.1